+++ src/main/java/shop/sajotuna/order/orders/repository/OrderProductRepository.java
@@ -3,6 +3,6 @@ package shop.sajotuna.order.orders.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import shop.sajotuna.order.orders.entity.OrderProduct;
 
-public interface OrderProductRepository extends JpaRepository<OrderProduct, Integer> {
+public interface OrderProductRepository extends JpaRepository<OrderProduct, Long> {
 
 }
